third_party/blink/web_tests/html/tabular_data/tr_insertcell_skips_non_td_th-expected.txt

Tests that HTMLTableRowElement.insertCell() skips non <td> / <th> children.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS tr.__proto__ is HTMLTableRowElement.prototype
PASS childNodes.length is 4
PASS childNodes[0].nodeValue is "TEXT"
PASS childNodes[1].tagName is "A"
PASS childNodes[2].innerHTML is "0"
PASS childNodes[3].innerHTML is "1"
PASS successfullyParsed is true

TEST COMPLETE
